+++ test/factories/uploads.rb
@@ -1,16 +1,23 @@
 FactoryGirl.define do
   factory :upload, class: "S3Relay::Upload" do
-    uuid         SecureRandom.uuid
+    uuid { SecureRandom.uuid }
+  end
+
+  factory :file_upload, parent: :upload do
     filename     "cat.png"
     content_type "image/png"
     upload_type  "FileUpload"
   end
 
   factory :icon_upload, parent: :upload do
-    upload_type "IconUpload"
+    filename     "cat.png"
+    content_type "image/png"
+    upload_type  "IconUpload"
   end
 
   factory :photo_upload, parent: :upload do
-    upload_type "PhotoUpload"
+    filename     "cat.png"
+    content_type "image/png"
+    upload_type  "PhotoUpload"
   end
 end